python-client/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=3.20.0 